completed blogs API

+++ src/main/java/com/evolv/blogCRUD/services/BlogServiceImpl.java
@@ -5,7 +5,10 @@ import com.evolv.blogCRUD.entities.Blog;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class BlogServiceImpl implements BlogService {
@@ -29,17 +32,41 @@ public class BlogServiceImpl implements BlogService {
     }
 
     @Override
-    public void deleteBlogById(long blogId) {
-        blogDao.deleteById(blogId);
+    public void deleteBlogById(long blogId) throws NoSuchElementException{
+        if(blogDao.existsById(blogId)) {
+            blogDao.deleteById(blogId);
+        }
+        else{
+            throw new NoSuchElementException();
+        }
     }
 
     @Override
-    public void updateBlogById(long blogId,Blog blog) throws Exception {
+    public void updateBlogById(long blogId,Blog blog) throws NoSuchElementException {
         if(blogDao.existsById(blogId)) {
-            blogDao.save(blog);
+            Blog b = blogDao.getOne(blogId);
+            b.setLastUpdated(new Date());
+            if(blog.getContent()!=null)
+            {
+                b.setContent(blog.getContent());
+            }
+
+            if(blog.getTitle()!=null)
+            {
+                b.setTitle(blog.getTitle());
+            }
+            if(blog.getSummary()!=null)
+            {
+                b.setSummary(blog.getSummary());
+            }
+            if(blog.getSlug()!=null)
+            {
+                b.setSlug(blog.getSlug());
+            }
+            blogDao.save(b);
         }
         else{
-            throw new Exception("No blog with given Id");
+            throw new NoSuchElementException();
         }
     }
 }
